+++ ng2-components/ng2-activiti-analytics/src/assets/analyticsComponent.mock.ts
@@ -15,109 +15,6 @@
  * limitations under the License.
  */
 
-import { ReportParameterDetailsModel } from '../models/report.model';
-
-export var reportDefParamStatus = {
-    'id': 2005,
-    'name': 'Fake Task overview status',
-    'created': '2016-10-05T15:39:40.222+0000',
-    'definition': '{ "parameters" :[{"id":"status","name":null,"nameKey":null,"type":"status","value":null,"dependsOn":null}]}'
-};
-
-export var reportDefParamNumber = {
-    'id': 2005,
-    'name': 'Fake Process instances overview',
-    'created': '2016-10-05T15:39:40.222+0000',
-    'definition': '{ "parameters"' +
-    ' :[{"id":"slowProcessInstanceInteger","name":null,"nameKey":null,"type":"integer","value":10,"dependsOn":null}]}'
-};
-
-export var reportDefParamDuration = {
-    'id': 2005,
-    'name': 'Fake Task service level agreement',
-    'created': '2016-10-05T15:39:40.222+0000',
-    'definition': '{ "parameters"' +
-    ' :[{"id":"duration","name":null,"nameKey":null,"type":"duration","value":null,"dependsOn":null}]}'
-};
-
-export var reportDefParamCheck = {
-    'id': 2005,
-    'name': 'Fake Task service level agreement',
-    'created': '2016-10-05T15:39:40.222+0000',
-    'definition': '{ "parameters"' +
-    ' :[{"id":"typeFiltering","name":null,"nameKey":null,"type":"boolean","value":true,"dependsOn":null}]}'
-};
-
-export var reportDefParamDateRange = {
-    'id': 2005,
-    'name': 'Fake Process instances overview',
-    'created': '2016-10-05T15:39:40.222+0000',
-    'definition': '{ "parameters" :[{"id":"dateRange","name":null,"nameKey":null,"type":"dateRange","value":null,"dependsOn":null}]}'
-};
-
-export var reportDefParamRangeInterval = {
-    'id': 2006,
-    'name': 'Fake Task overview RangeInterval',
-    'created': '2016-10-05T15:39:40.222+0000',
-    'definition': '{ "parameters" :[{"id":"dateRangeInterval","name":null,"nameKey":null,"type":"dateInterval","value":null,"dependsOn":null}]}'
-};
-
-export var reportDefParamProcessDef = {
-    'id': 2006,
-    'name': 'Fake Task overview ProcessDefinition',
-    'created': '2016-10-05T15:39:40.222+0000',
-    'definition': '{ "parameters" :[{"id":"processDefinitionId","name":null,"nameKey":null,"type":"processDefinition","value":null,"dependsOn":null}]}'
-};
-
-export var reportDefParamProcessDefOptions = {
-    'size': 4, 'total': 4, 'start': 0, 'data': [
-        {
-            'id': 'FakeProcessTest 1:1:1',
-            'name': 'Fake Process Test 1 Name ',
-            'version': 1
-        },
-        {
-            'id': 'FakeProcessTest 1:2:1',
-            'name': 'Fake Process Test 1 Name ',
-            'version': 2
-        },
-        {
-            'id': 'FakeProcessTest 2:1:1',
-            'name': 'Fake Process Test 2 Name ',
-            'version': 1
-        },
-        {
-            'id': 'FakeProcessTest 3:1:1',
-            'name': 'Fake Process Test 3 Name ',
-            'version': 1
-        }
-    ]
-};
-
-export var reportDefParamProcessDefOptionsApp = {
-    'size': 2, 'total': 2, 'start': 2, 'data': [
-        {
-            'id': 'FakeProcessTest 1:1:1',
-            'name': 'Fake Process Test 1 Name ',
-            'version': 1
-        },
-        {
-            'id': 'FakeProcessTest 1:2:1',
-            'name': 'Fake Process Test 1 Name ',
-            'version': 2
-        }
-    ]
-};
-
-export var reportDefParamTask = {
-    'id': 2006,
-    'name': 'Fake Task service level agreement',
-    'created': '2016-10-05T15:39:40.222+0000',
-    'definition': '{ "parameters" :[{"id":"taskName","name":null,"nameKey":null,"type":"task","value":null,"dependsOn":"processDefinitionId"}]}'
-};
-
-export var reportDefParamTaskOptions = ['Fake task name 1', 'Fake task name 2'];
-
 export var chartProcessDefOverview = {
     'elements': [{
         'id': 'id1585876275153',
@@ -198,58 +95,3 @@ export var chartTaskOverview = {
         ]
     }]
 };
-
-export var fieldNumber = new ReportParameterDetailsModel(
-    {
-        id: 'slowProcessInstanceInteger',
-        type: 'integer',
-        value: '102'
-    }
-);
-
-export var fieldStatus = new ReportParameterDetailsModel(
-    {
-        id: 'status',
-        type: 'status',
-        value: 'fake-value'
-    }
-);
-
-export var fieldTypeFiltering = new ReportParameterDetailsModel(
-    {
-        id: 'typeFiltering',
-        type: 'boolean',
-        value: false
-    }
-);
-
-export var fieldTask = new ReportParameterDetailsModel(
-    {
-        id: 'taskName',
-        type: 'task',
-        value: 'fake-task-name'
-    }
-);
-
-export var fieldDateRange = {
-    startDate: '2016-10-12T00:00:00.000Z',
-    endDate: '2016-10-14T00:00:00.000Z'
-};
-
-export var fieldDateRangeInterval = new ReportParameterDetailsModel(
-    {
-        id: 'dateRangeInterval',
-        type: 'dateInterval',
-        value: 'fake-date-interval'
-    }
-);
-
-export var fieldProcessDef = new ReportParameterDetailsModel(
-    {
-        id: 'processDefinitionId',
-        type: 'processDefinition',
-        value: 'fake-process-name:1:15027'
-    }
-);
-
-export var fieldDuration = {value: 30};
